requirements.txt
rumps>=0.4.0
pyobjc-framework-Cocoa>=9.0
openai>=1.0.0
pillow>=10.0.0
